+++ include/opengl/libqopenglfunctions_3_2_compatibility.h
@@ -768,28 +768,28 @@ void* q_openglfunctions_3_2_compatibility_gl_map_buffer(void* self, uint32_t tar
 ///
 /// @param self QOpenGLFunctions_3_2_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param size int64_t
+/// @param offset intptr_t
+/// @param size intptr_t
 /// @param data void*
-void q_openglfunctions_3_2_compatibility_gl_get_buffer_sub_data(void* self, uint32_t target, int64_t offset, int64_t size, void* data);
+void q_openglfunctions_3_2_compatibility_gl_get_buffer_sub_data(void* self, uint32_t target, intptr_t offset, intptr_t size, void* data);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glBufferSubData)
 ///
 /// @param self QOpenGLFunctions_3_2_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param size int64_t
+/// @param offset intptr_t
+/// @param size intptr_t
 /// @param data void*
-void q_openglfunctions_3_2_compatibility_gl_buffer_sub_data(void* self, uint32_t target, int64_t offset, int64_t size, void* data);
+void q_openglfunctions_3_2_compatibility_gl_buffer_sub_data(void* self, uint32_t target, intptr_t offset, intptr_t size, void* data);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glBufferData)
 ///
 /// @param self QOpenGLFunctions_3_2_Compatibility*
 /// @param target uint32_t
-/// @param size int64_t
+/// @param size intptr_t
 /// @param data void*
 /// @param usage uint32_t
-void q_openglfunctions_3_2_compatibility_gl_buffer_data(void* self, uint32_t target, int64_t size, void* data, uint32_t usage);
+void q_openglfunctions_3_2_compatibility_gl_buffer_data(void* self, uint32_t target, intptr_t size, void* data, uint32_t usage);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glIsBuffer)
 ///
@@ -1360,18 +1360,18 @@ void q_openglfunctions_3_2_compatibility_gl_bind_vertex_array(void* self, uint32
 ///
 /// @param self QOpenGLFunctions_3_2_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param length int64_t
-void q_openglfunctions_3_2_compatibility_gl_flush_mapped_buffer_range(void* self, uint32_t target, int64_t offset, int64_t length);
+/// @param offset intptr_t
+/// @param length intptr_t
+void q_openglfunctions_3_2_compatibility_gl_flush_mapped_buffer_range(void* self, uint32_t target, intptr_t offset, intptr_t length);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glMapBufferRange)
 ///
 /// @param self QOpenGLFunctions_3_2_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param length int64_t
+/// @param offset intptr_t
+/// @param length intptr_t
 /// @param access uint32_t
-void* q_openglfunctions_3_2_compatibility_gl_map_buffer_range(void* self, uint32_t target, int64_t offset, int64_t length, uint32_t access);
+void* q_openglfunctions_3_2_compatibility_gl_map_buffer_range(void* self, uint32_t target, intptr_t offset, intptr_t length, uint32_t access);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glFramebufferTextureLayer)
 ///
@@ -1754,9 +1754,9 @@ void q_openglfunctions_3_2_compatibility_gl_bind_buffer_base(void* self, uint32_
 /// @param target uint32_t
 /// @param index uint32_t
 /// @param buffer uint32_t
-/// @param offset int64_t
-/// @param size int64_t
-void q_openglfunctions_3_2_compatibility_gl_bind_buffer_range(void* self, uint32_t target, uint32_t index, uint32_t buffer, int64_t offset, int64_t size);
+/// @param offset intptr_t
+/// @param size intptr_t
+void q_openglfunctions_3_2_compatibility_gl_bind_buffer_range(void* self, uint32_t target, uint32_t index, uint32_t buffer, intptr_t offset, intptr_t size);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glEndTransformFeedback)
 ///
@@ -1821,10 +1821,10 @@ void q_openglfunctions_3_2_compatibility_gl_color_maski(void* self, uint32_t ind
 /// @param self QOpenGLFunctions_3_2_Compatibility*
 /// @param readTarget uint32_t
 /// @param writeTarget uint32_t
-/// @param readOffset int64_t
-/// @param writeOffset int64_t
-/// @param size int64_t
-void q_openglfunctions_3_2_compatibility_gl_copy_buffer_sub_data(void* self, uint32_t readTarget, uint32_t writeTarget, int64_t readOffset, int64_t writeOffset, int64_t size);
+/// @param readOffset intptr_t
+/// @param writeOffset intptr_t
+/// @param size intptr_t
+void q_openglfunctions_3_2_compatibility_gl_copy_buffer_sub_data(void* self, uint32_t readTarget, uint32_t writeTarget, intptr_t readOffset, intptr_t writeOffset, intptr_t size);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-3-2-compatibility.html#glUniformBlockBinding)
 ///
