+++ include/opengl/libqopenglfunctions_4_3_compatibility.h
@@ -768,28 +768,28 @@ void* q_openglfunctions_4_3_compatibility_gl_map_buffer(void* self, uint32_t tar
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param size int64_t
+/// @param offset intptr_t
+/// @param size intptr_t
 /// @param data void*
-void q_openglfunctions_4_3_compatibility_gl_get_buffer_sub_data(void* self, uint32_t target, int64_t offset, int64_t size, void* data);
+void q_openglfunctions_4_3_compatibility_gl_get_buffer_sub_data(void* self, uint32_t target, intptr_t offset, intptr_t size, void* data);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glBufferSubData)
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param size int64_t
+/// @param offset intptr_t
+/// @param size intptr_t
 /// @param data void*
-void q_openglfunctions_4_3_compatibility_gl_buffer_sub_data(void* self, uint32_t target, int64_t offset, int64_t size, void* data);
+void q_openglfunctions_4_3_compatibility_gl_buffer_sub_data(void* self, uint32_t target, intptr_t offset, intptr_t size, void* data);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glBufferData)
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param target uint32_t
-/// @param size int64_t
+/// @param size intptr_t
 /// @param data void*
 /// @param usage uint32_t
-void q_openglfunctions_4_3_compatibility_gl_buffer_data(void* self, uint32_t target, int64_t size, void* data, uint32_t usage);
+void q_openglfunctions_4_3_compatibility_gl_buffer_data(void* self, uint32_t target, intptr_t size, void* data, uint32_t usage);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glIsBuffer)
 ///
@@ -1360,18 +1360,18 @@ void q_openglfunctions_4_3_compatibility_gl_bind_vertex_array(void* self, uint32
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param length int64_t
-void q_openglfunctions_4_3_compatibility_gl_flush_mapped_buffer_range(void* self, uint32_t target, int64_t offset, int64_t length);
+/// @param offset intptr_t
+/// @param length intptr_t
+void q_openglfunctions_4_3_compatibility_gl_flush_mapped_buffer_range(void* self, uint32_t target, intptr_t offset, intptr_t length);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glMapBufferRange)
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param target uint32_t
-/// @param offset int64_t
-/// @param length int64_t
+/// @param offset intptr_t
+/// @param length intptr_t
 /// @param access uint32_t
-void* q_openglfunctions_4_3_compatibility_gl_map_buffer_range(void* self, uint32_t target, int64_t offset, int64_t length, uint32_t access);
+void* q_openglfunctions_4_3_compatibility_gl_map_buffer_range(void* self, uint32_t target, intptr_t offset, intptr_t length, uint32_t access);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glFramebufferTextureLayer)
 ///
@@ -1754,9 +1754,9 @@ void q_openglfunctions_4_3_compatibility_gl_bind_buffer_base(void* self, uint32_
 /// @param target uint32_t
 /// @param index uint32_t
 /// @param buffer uint32_t
-/// @param offset int64_t
-/// @param size int64_t
-void q_openglfunctions_4_3_compatibility_gl_bind_buffer_range(void* self, uint32_t target, uint32_t index, uint32_t buffer, int64_t offset, int64_t size);
+/// @param offset intptr_t
+/// @param size intptr_t
+void q_openglfunctions_4_3_compatibility_gl_bind_buffer_range(void* self, uint32_t target, uint32_t index, uint32_t buffer, intptr_t offset, intptr_t size);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glEndTransformFeedback)
 ///
@@ -1821,10 +1821,10 @@ void q_openglfunctions_4_3_compatibility_gl_color_maski(void* self, uint32_t ind
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param readTarget uint32_t
 /// @param writeTarget uint32_t
-/// @param readOffset int64_t
-/// @param writeOffset int64_t
-/// @param size int64_t
-void q_openglfunctions_4_3_compatibility_gl_copy_buffer_sub_data(void* self, uint32_t readTarget, uint32_t writeTarget, int64_t readOffset, int64_t writeOffset, int64_t size);
+/// @param readOffset intptr_t
+/// @param writeOffset intptr_t
+/// @param size intptr_t
+void q_openglfunctions_4_3_compatibility_gl_copy_buffer_sub_data(void* self, uint32_t readTarget, uint32_t writeTarget, intptr_t readOffset, intptr_t writeOffset, intptr_t size);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glUniformBlockBinding)
 ///
@@ -3739,9 +3739,9 @@ void q_openglfunctions_4_3_compatibility_gl_tex_storage2_d_multisample(void* sel
 /// @param target uint32_t
 /// @param internalformat uint32_t
 /// @param buffer uint32_t
-/// @param offset int64_t
-/// @param size int64_t
-void q_openglfunctions_4_3_compatibility_gl_tex_buffer_range(void* self, uint32_t target, uint32_t internalformat, uint32_t buffer, int64_t offset, int64_t size);
+/// @param offset intptr_t
+/// @param size intptr_t
+void q_openglfunctions_4_3_compatibility_gl_tex_buffer_range(void* self, uint32_t target, uint32_t internalformat, uint32_t buffer, intptr_t offset, intptr_t size);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glShaderStorageBlockBinding)
 ///
@@ -3824,9 +3824,9 @@ void q_openglfunctions_4_3_compatibility_gl_invalidate_buffer_data(void* self, u
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param buffer uint32_t
-/// @param offset int64_t
-/// @param length int64_t
-void q_openglfunctions_4_3_compatibility_gl_invalidate_buffer_sub_data(void* self, uint32_t buffer, int64_t offset, int64_t length);
+/// @param offset intptr_t
+/// @param length intptr_t
+void q_openglfunctions_4_3_compatibility_gl_invalidate_buffer_sub_data(void* self, uint32_t buffer, intptr_t offset, intptr_t length);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glInvalidateTexImage)
 ///
@@ -3921,9 +3921,9 @@ void q_openglfunctions_4_3_compatibility_gl_vertex_attrib_format(void* self, uin
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param bindingindex uint32_t
 /// @param buffer uint32_t
-/// @param offset int64_t
+/// @param offset intptr_t
 /// @param stride int32_t
-void q_openglfunctions_4_3_compatibility_gl_bind_vertex_buffer(void* self, uint32_t bindingindex, uint32_t buffer, int64_t offset, int32_t stride);
+void q_openglfunctions_4_3_compatibility_gl_bind_vertex_buffer(void* self, uint32_t bindingindex, uint32_t buffer, intptr_t offset, int32_t stride);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glTextureView)
 ///
@@ -3961,8 +3961,8 @@ void q_openglfunctions_4_3_compatibility_gl_copy_image_sub_data(void* self, uint
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glDispatchComputeIndirect)
 ///
 /// @param self QOpenGLFunctions_4_3_Compatibility*
-/// @param indirect int64_t
-void q_openglfunctions_4_3_compatibility_gl_dispatch_compute_indirect(void* self, int64_t indirect);
+/// @param indirect intptr_t
+void q_openglfunctions_4_3_compatibility_gl_dispatch_compute_indirect(void* self, intptr_t indirect);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glDispatchCompute)
 ///
@@ -3977,12 +3977,12 @@ void q_openglfunctions_4_3_compatibility_gl_dispatch_compute(void* self, uint32_
 /// @param self QOpenGLFunctions_4_3_Compatibility*
 /// @param target uint32_t
 /// @param internalformat uint32_t
-/// @param offset int64_t
-/// @param size int64_t
+/// @param offset intptr_t
+/// @param size intptr_t
 /// @param format uint32_t
 /// @param type uint32_t
 /// @param data void*
-void q_openglfunctions_4_3_compatibility_gl_clear_buffer_sub_data(void* self, uint32_t target, uint32_t internalformat, int64_t offset, int64_t size, uint32_t format, uint32_t type, void* data);
+void q_openglfunctions_4_3_compatibility_gl_clear_buffer_sub_data(void* self, uint32_t target, uint32_t internalformat, intptr_t offset, intptr_t size, uint32_t format, uint32_t type, void* data);
 
 /// [Qt documentation](https://doc.qt.io/qt-6/qopenglfunctions-4-3-compatibility.html#glClearBufferData)
 ///
